Terminate request should be sent only in active state

+++ app/assets/javascripts/interaction_web_tools/chat.coffee
@@ -94,9 +94,10 @@ class InteractionWebTools.Chat.Client
     @autoScroll()
 
   terminateChat: =>
-    $.ajax
-      url: @endpoint
-      type: 'DELETE'
+    if @state == stateEnum.ACTIVE
+      $.ajax
+        url: @endpoint
+        type: 'DELETE'
     @state = stateEnum.TERMINATED
 
   displayIndicator: (type) =>
